% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Allclasses.R
\name{catalog,tempoSig-method}
\alias{catalog,tempoSig-method}
\title{Accessor for catalog}
\usage{
\S4method{catalog}{tempoSig}(object)
}
\arguments{
\item{object}{Object containing catalog}
}
\value{
\code{catalog} data frame
}
\description{
Accessor for catalog
}
