% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refitting.R
\name{CosineSimilarity}
\alias{CosineSimilarity}
\title{Cosine similarity}
\usage{
CosineSimilarity(A, B)
}
\arguments{
\item{A}{Test matrix of dimension \code{(m,a)}}

\item{B}{Reference matrix of dimension \code{(m, b)}}
}
\value{
Matrix of dimension \code{(a,b)}, whose elements give overlap of column \code{a}
        in matrix \code{A} with column \code{b} in matrix \code{B}.
}
\description{
Cosine similarity
}
