% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities3.R
\name{filterExposure}
\alias{filterExposure}
\title{Filter exposures using p-values}
\usage{
filterExposure(object, alpha = 0.05, attribution = TRUE)
}
\arguments{
\item{object}{Object of class \code{tempoSig}}

\item{alpha}{Type-I error treshold}

\item{attribution}{If \code{TRUE}, output is multiplied by mutation counts of each sample}
}
\value{
Matrix of filtered exposure or attribution
}
\description{
Filter exposures using p-values
}
