% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities2.R
\name{filterFP}
\alias{filterFP}
\title{Filter False Positives}
\usage{
filterFP(exposure, pval, alpha = 0.05, min.M = NA)
}
\arguments{
\item{exposure}{Exposure input data file}

\item{pval}{P-value input data file}

\item{alpha}{FPR threshold}

\item{min.M}{Minimum mutation count}
}
\value{
Data frame of filtered exposure data
}
\description{
Use exposure and p-value tables and generated filtered exposures
}
