% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{fitMLE}
\alias{fitMLE}
\title{Maximum likelihood inference of signature proportions}
\usage{
fitMLE(x, ref, itmax = 1000, tol = 1e-04, kfixed = -1)
}
\arguments{
\item{x}{Vector of observed proportions of mutation contexts}

\item{ref}{Matrix of reference signature proportions with mutation types in rows
and signatures in columns}

\item{itmax}{Maximum number of iterations}

\item{tol}{Tolerance of convergence}

\item{kfixed}{Index of parameter to be constrained as zero}
}
\value{
List of exposure vector \code{p} and \code{loglik}
}
\description{
Maximum likelihood inference of signature proportions
}
