% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Allclasses.R
\name{logLik,tempoSig-method}
\alias{logLik,tempoSig-method}
\title{Accessor for logLik}
\usage{
\S4method{logLik}{tempoSig}(object)
}
\arguments{
\item{object}{Object containing \code{logLik}}
}
\value{
\code{logLik} value
}
\description{
Accessor for logLik
}
