% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refitting.R
\name{mutationalCone}
\alias{mutationalCone}
\title{Refit spectra using MutationalCone}
\usage{
mutationalCone(catalog, signatures = "COSMIC.v2", normalize = FALSE)
}
\arguments{
\item{catalog}{Input matrix of dimension \code{(M, N)}, where
\code{M} is the no. of genetic features, e.g., 96 for SNV trinucleotides)
and \code{N} is the number of samples to refit}

\item{signatures}{Matrix of dimension \code{(M,K)} with \code{K} reference 
signatures in columns (e.g., COSMIC signatures).}

\item{normalize}{Normalize output column-wise}
}
\value{
Matrix of dimension \code{(K, N)} with columns giving estimated
        loading of signatures in each sample.
}
\description{
Given input spectra, find proportions of known signatures
}
\details{
Adopted from Omichessan et al. 
\url{http://dx.doi.org/10.1371/journal.pone.0221235}.
}
