% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgnmf.R
\name{pgnmf}
\alias{pgnmf}
\title{piggyback NMF}
\usage{
pgnmf(x, stock = NULL, filter = FALSE, cutoff = NULL, progress.bar = TRUE, ...)
}
\arguments{
\item{x}{Catalog matrix to be analyzed}

\item{stock}{\code{tempoSig} object containing the stock solution}

\item{filter}{Filter exposure using CV cutoff \code{cutoff}}

\item{cutoff}{CV cutoff matrix for filtering}
}
\value{
List of \code{mean} and coefficient of variation \code{cv} (sd/mean) of exposure matrices
}
\description{
Enables de novo inference of single sample using pre-inferred exposure 
and signature sets for a large data set (stock)
}
