% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{plotExposure}
\alias{plotExposure}
\title{Plot exposure}
\usage{
plotExposure(object, sample.id, cutoff = 0.001, ...)
}
\arguments{
\item{object}{Object of class \code{tempoSig}}

\item{sample.id}{Sample index or name (\code{Tumor_Sample_Barcode}) to display}

\item{cutoff}{Minimum proportion for displaying signature labels}

\item{...}{Other parameters to \code{barplot}}
}
\value{
None
}
\description{
Display barplot of exposure extracted for a specific sample
}
