% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epower.R
\name{powerEstimate}
\alias{powerEstimate}
\title{Estimate Power for Signature Detection}
\usage{
powerEstimate(
  proportion,
  reference = "v2",
  use.model = TRUE,
  nsamp = 100,
  Mcoeff = NULL,
  afit = NULL,
  bfit = NULL,
  Msamp = NULL,
  verbose = TRUE,
  alpha = 0.05,
  Pmin = 0.8,
  ...
)
}
\arguments{
\item{proportion}{Vector of signature proportions}

\item{reference}{Reference signature version; currently only supports \code{c('SA')}.}

\item{use.model}{Use parametrized model. If \code{FALSE}, multinomial model will be simulated.}

\item{nsamp}{No. of samples to use for simulation}

\item{Msamp}{Vector mutation loads to sample}

\item{verbose}{Print messages during sampling}

\item{alpha}{False positive rate}

\item{Pmin}{Lower bound for power}

\item{cut}{Remove signatures with proportions below \code{cut}.}
}
\description{
Estimate Power for Signature Detection
}
