% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Allclasses.R
\name{pvalue,tempoSig-method}
\alias{pvalue,tempoSig-method}
\title{Accessor for p-values}
\usage{
\S4method{pvalue}{tempoSig}(object)
}
\arguments{
\item{object}{Object containing \code{pvalue}}
}
\value{
\code{pvalue} data frame
}
\description{
Accessor for p-values
}
