% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities3.R
\name{readMAFforcat}
\alias{readMAFforcat}
\title{Read MAF file and Fix Chromosome Names}
\usage{
readMAFforcat(maf)
}
\arguments{
\item{maf}{MAF file path}
}
\value{
Data frame with chromosome names fixed
}
\description{
This function is a wrapper intended for use in \code{maf2cat3}, which requires chromosome names
compatible with \code{BSgenome.Hsapiens.UCSC.hg19} (\code{chr1, ..., chrX, chrY}).
}
