% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra.R
\name{Row2Column}
\alias{Row2Column}
\title{Reformat spectra to column-major}
\usage{
Row2Column(X)
}
\description{
Transform spectral matrix with mutation contexts in rows to its transpose
with columns renamed for Mutation-signatures input
}
