% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{run_tSig}
\alias{run_tSig}
\title{Wrapper Function}
\usage{
run_tSig(
  catalog,
  signat = "v2",
  nperm = 1000,
  alpha = 0.05,
  output = "exposure.txt"
)
}
\arguments{
\item{catalog}{Catalog input file}

\item{signat}{Reference signature choices, \code{c('v2','SA','SP')} for sCOSMIC v2, 
SigAnalyzer, and sigProfiler, respectively}

\item{nperm}{No. of permutations}

\item{alpha}{False positive rate threshold}

\item{output}{Output file}
}
\description{
One-step execution of exposure inference, filtering, and output
}
