% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{senspec}
\alias{senspec}
\title{True and false positive rates}
\usage{
senspec(xhat, x, pvalue = NULL, alpha = 0.05)
}
\arguments{
\item{xhat}{Predicted exposure vector}

\item{x}{True exposure vector}

\item{pvalue}{Vector of prediction p-values}

\item{alpha}{False positive rate threshold}
}
\value{
list of vectors \code{tp} (true positives), \code{fp} (false positives),
        all positives \code{positives} and negatives \code{negatives} in
        names of \code{xhat}
}
\description{
From supplied exposure prediction and actual vectors, compute 
true and false positive rates
}
