% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Allclasses.R
\name{show,tempoSig-method}
\alias{show,tempoSig-method}
\title{Display \code{tempoSig} object}
\usage{
\S4method{show}{tempoSig}(object)
}
\arguments{
\item{object}{Object of class \code{tempoSig}}
}
\value{
\code{NULL}
}
\description{
Display the class and dimension
}
