% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra.R
\name{simulateSpectra}
\alias{simulateSpectra}
\title{Simulate mutational spectra}
\usage{
simulateSpectra(
  W,
  pzero = 0.5,
  nmut = 100,
  h,
  N = 10,
  dilute.ultra = FALSE,
  min.mut = 1
)
}
\arguments{
\item{W}{Input signature matrix}

\item{pzero}{Proportion of extra zero counts (scalar)}

\item{nmut}{Mean number of mutations per sample (scalar)}

\item{h}{Vector of average proportion of mutations in each process;
length equal to the number of columns in \code{W}.}

\item{N}{No. of samples}

\item{dilute.ultra}{Dipute ultra-mutated samples to reduce bias}

\item{min.mut}{Minimum mutation load}
}
\value{
List of components \code{X}: simulated mutation counts;
        \code{W}, \code{H}: signature and exposure matrices.
}
\description{
Generate simulated data using zero-inflated Poisson model
}
\details{
See Omichessan et al. DOI: 10.1371/journal.pone.0221235
}
\examples{
library(gtools)
set.seed(135)
K <- 5 # no. of processes
W <- t(rdirichlet(n = K, alpha = rep(1, 96)))
h <- rdirichlet(n = K, alpha = rep(5, K))[1,]
x <- simulateSpectra(W = W, h = h, N = 100)
}
