% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Allclasses.R, R/tempoSig.R
\docType{package}
\name{tempoSig}
\alias{tempoSig}
\title{Create \code{tempoSig} object}
\usage{
tempoSig(data, signat = "SP")
}
\arguments{
\item{data}{Input data of mutation catalog; samples in columns
and mutation types in rows.}

\item{signatures}{Reference signatures table; default is COSMIC
version 3 signatures.}
}
\value{
Object of class \code{tempoSig}.
}
\examples{
set.seed(130)
data <- rmultinom(n = 5, size = 100, prob = rep(1,96))
rownames(data) <- trinucleotides()
colnames(data) <- paste0('S',seq(5))
s <- tempoSig(data)
data <- read.table(system.file('extdata', 'tcga-brca_catalog.txt',
                  package='tempoSig'), header=TRUE, sep='\t')
b <- tempoSig(data)
}
