% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refitting.R
\name{cosineSimilarity}
\alias{cosineSimilarity}
\title{Cosine similarity}
\usage{
cosineSimilarity(A, B, diag = FALSE)
}
\arguments{
\item{A}{Test matrix of dimension \code{(m,a)}}

\item{B}{Reference matrix of dimension \code{(m, b)}}

\item{diag}{Only compare column \code{A[, i]} with column \code{B[, i]} 
where \code{i=1, ..., ncol(A)=ncol(B)}.}
}
\value{
If \code{diag = TRUE}, vector of overlap between columns of \code{A}
        and columns of \code{B} in one-to-one mapping; if \code{diag = FALSE},
        matrix of dimension \code{(a,b)}, whose elements give overlap of column 
        \code{a} in matrix \code{A} with column \code{b} in matrix \code{B}.
}
\description{
Row names of two data matrices (mutation contexts) are used for comparison.
If the row names do not match completely, the larger set is used as reference 
with the smaller set expanded with zero-padding.
}
