% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{fitMLE}
\alias{fitMLE}
\title{Maximum likelihood inference of signature proportions}
\usage{
fitMLE(x, ref, itmax = 1000, tol = 1e-04, min.tmb = 5)
}
\arguments{
\item{x}{Vector of observed proportions of mutation contexts}

\item{ref}{Matrix of reference signature proportions with mutation types in rows
and signatures in columns}

\item{itmax}{Maximum number of iterations}

\item{tol}{Tolerance of convergence}

\item{min.tmb}{Minimum mutation counts; if lower than this, \code{NA} is returned}
}
\value{
Vector of estimated signature exposure proportions
}
\description{
Maximum likelihood inference of signature proportions
}
