% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{run_tSig}
\alias{run_tSig}
\title{Wrapper Function}
\usage{
run_tSig(
  catalog,
  signat = "SA",
  nperm = 1000,
  alpha = 0.05,
  output = "exposure.txt"
)
}
\arguments{
\item{catalog}{Catalog input file}

\item{signat}{Reference signature choices, \code{c('SA','SP', 'cosmic2')} for sigAnalyzer, sigProfiler, and
COSMIC version 2, respectively}

\item{nperm}{No. of permutations}

\item{alpha}{False positive rate threshold}

\item{output}{Output file}
}
\description{
One-step execution of exposure inference, filtering, and outpiut
}
