% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities3.R
\name{sigplot}
\alias{sigplot}
\title{Plot signature profile}
\usage{
sigplot(x, ...)
}
\arguments{
\item{x}{Signature profile vector; must be of length 96 and named for nucleotide contexts}

\item{...}{Other parameters for \code{plot}}
}
\value{
Return vector of \code{barplot}
}
\description{
Plot signature profile
}
